TRS provider now uses a functional id for authentication
Added two-legged OAuth support back into CredentialsFilter.java, its needed to be able to add the TRS provider as an LQE data source. The trs_user in the config.properties file is used as the functional ID for two-legged OAuth.

The TRS provider now logs in with the trs_user's credentials to access the IoT Platform users, and logs out once the TRS resource scan is completed.